requirements.txt
altgraph==0.17.4
anyio==4.8.0
APScheduler==3.11.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
coverage==7.7.0
crashtest==0.4.1
cryptography==42.0.5
cssselect2==0.7.0
curl_cffi==0.10.0
distlib==0.3.9
Django==5.0.4
django-ckeditor==6.7.1
django-cors-headers==4.7.0
django-crispy-forms==2.1
django-htmx==1.22.0
django-js-asset==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
dulwich==0.22.7
exceptiongroup==1.2.2
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.2
flake8==7.1.2
fonttools==4.51.0
frozendict==2.4.6
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
identify==2.6.8
idna==3.7
importlib_metadata==8.6.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
keyring==25.6.0
lxml==5.2.1
mccabe==0.7.0
more-itertools==10.6.0
msgpack==1.1.0
multitasking==0.0.11
nodeenv==1.9.1
numpy==2.2.5
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pbs-installer==2025.2.12
peewee==3.18.1
pefile==2023.2.7
pillow==10.3.0
pkginfo==1.12.1.2
platformdirs==4.3.6
poetry==2.1.1
poetry-core==2.1.1
protobuf>=3.20.3,<5.0.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pydyf==0.9.0
pyflakes==3.2.0
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pypdf==4.2.0
pyphen==0.14.0
pypng==0.20220715.0
pyproject_hooks==1.2.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.1
qrcode==7.4.2
RapidFuzz==3.12.1
redis==6.1.0
reportlab==4.0.9
requests==2.31.0
requests-toolbelt==1.0.0
ruff==0.11.12
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.2.1
tomli==2.2.1
tomlkit==0.13.2
trove-classifiers==2025.2.18.16
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
uvicorn==0.34.2
virtualenv==20.29.2
weasyprint==61.2
webencodings==0.5.1
websockets==15.0.1
whitenoise==6.9.0
xhtml2pdf==0.2.15
yfinance==0.2.61
zipp==3.21.0
zopfli==0.2.3
zstandard==0.23.0
gunicorn==21.2.0
google-generativeai
resend==2.11.0
django-unfold==0.63.0
django-money==3.5.4
djlint
weasyprint
python-json-logger
django-ratelimit
django-axes 
django-csp
sentry-sdk
daphne
django-ipware